Write a function to create an aggregate score for a place
Function takes creates a % score for a rating source via this formula
source_score = (score - 1) / (source_max_rating - 1)
Why the -1's? Most review sources when asking for a rating ask for a 
rating of 1-5. So a business with 1 star has only gotten the worst 
possible reviews. Because the aggregate score is out of 100, people 
might assume there are places in the 0-20 range. However, if the 1-5 
formula is used for calculation, it would be impossible to go under 20. 
Therefore ratings are measured on a scale of 0 to (MAX - 1) to keep 
ratings honest.