Add more undefined checks and make code run with undefined fields
Several functions would crash if even one field was missing from 
results, which is likely to happen from time to time. So every variable 
is checked to see if it exists, and given an empty value if it doesn't 
so the handling functions can identify missing data and act accordingly. 
After this commit, it will run over the sample data without error!