Add functions to rate places by city and fix yelp address madness
Include a few functions to expand the ones from the last commit. 
Basically discoverFood() takes a city, state, and optional text query 
and searches for restaurants in that location, then gets their ratings 
and dumps them to a spreadsheet. This uses a lot more data than any 
previous functions, and as such getAddressComponents() has been dealing 
with a lot of annoying edge cases. It has recently been discovered that 
yelp doesn't support everywhere, so there will be an array of country 
codes to check before running yelp ratings