Switch from UrlFetchApp to fetch() API. Make tests asynchronous
This program started development as a Google Web App, the developer did 
not know this when they did it. After deciding to build a frontend they 
learned that UrlFetchApp is server-side code that doesn't work from the 
client. Now all functions use the Fetch API and are called by 
asynchronous functions that will await the API response before returning 
a value;